Add CLI to take/release I2C mutex (#154)
* add semaphore control task
* create SoftwareCommands.{c,h} -- Move a bunch more commands out of CommandLineTask.c
* remove some obsolete print statements in MonitorTask and MonitorI2CTask
* limit task names to 5 chars
* cleanup of old-style prototypes continues
