+++ projects/cm_mcu/FreeRTOSConfig.h
@@ -93,9 +93,9 @@ to exclude the API function. */
 #define INCLUDE_vTaskDelay                  1
 #define INCLUDE_eTaskGetState               0
 #define INCLUDE_xTimerPendFunctionCall      0
-#define INCLUDE_xSemaphoreGetMutexHolder    0
+#define INCLUDE_xSemaphoreGetMutexHolder    1
 #define INCLUDE_xTaskGetHandle              0
-#define INCLUDE_xTaskGetCurrentTaskHandle   0
+#define INCLUDE_xTaskGetCurrentTaskHandle   1
 #define INCLUDE_xTaskGetIdleTaskHandle      0
 #define INCLUDE_xTaskAbortDelay             0
 #define INCLUDE_xTaskGetSchedulerState      0
