+++ projects/cm_mcu/InterruptHandlers.c
@@ -254,7 +254,6 @@ void SMBusMasterIntHandler5(void)
   portYIELD_FROM_ISR(xHigherPriorityTaskWoken);
 }
 
-
 void SMBusMasterIntHandler6(void)
 {
   BaseType_t xHigherPriorityTaskWoken = pdFALSE;
@@ -271,7 +270,7 @@ void SMBusMasterIntHandler6(void)
 // ADC conversion is complete.
 TaskHandle_t TaskNotifyADC = NULL;
 
-void ADCSeq0Interrupt()
+void ADCSeq0Interrupt(void)
 {
   BaseType_t xHigherPriorityTaskWoken = pdFALSE;
 
@@ -295,7 +294,7 @@ void ADCSeq0Interrupt()
   return;
 }
 
-void ADCSeq1Interrupt()
+void ADCSeq1Interrupt(void)
 {
   BaseType_t xHigherPriorityTaskWoken = pdFALSE;
 
@@ -322,7 +321,7 @@ void ADCSeq1Interrupt()
 // -----------------------------------------
 TaskHandle_t TaskNotifyI2CSlave = NULL;
 
-void I2CSlave0Interrupt()
+void I2CSlave0Interrupt(void)
 {
   BaseType_t xHigherPriorityTaskWoken = pdFALSE;
